pyproject.toml
# ~~ Generated by projen. To modify, edit .projenrc.py and run "npx projen".

[build-system]
requires = [ "poetry_core>=1.0.0" ]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "aws-rip"
version = "0.1.0"
description = ""
authors = [ "William Cleghorn <cleghornw@protonmail.com>" ]
readme = "README.md"

  [tool.poetry.dependencies]
  aws-cdk-lib = "^2.1.0"
  constructs = "^10.0.5"
  python = "^3.11"

  [tool.poetry.dev-dependencies]
  projen = "0.71.120"
